Ensure READY messages sent out by exiting worker are consumed prior to it actually exiting (issue #278) (#279)
* Ensure READY messages sent out via `Worker.workloop` are consumed

This commit fixes a race-condition which can lead to `WorkerLostError: Worker exited prematurely: exitcode 155` and `WorkerLostError: Worker exited prematurely: exitcode 0` errors.

The race-condition occurs when a message sent by `Worker.workloop` is not consumed by `ResultHandler` prior to the worker being cleaned up by `Pool._join_exited_workers`.

The solution to the race-condition is to keep a count of messages `ResultHandler` receives from a worker, and for the worker to return only when this count matches its own counter of completed tasks. This ensures that a worker will only be cleaned up once `ResultHandler` and the `Worker` are in-sync, or after 30 seconds of waiting for them to be in-sync.

* Remove accidental f-string usage

* Ensure "_ensure_messages_consumed" is always called in Worker.workloop

The previous change did not account for "wait_for_job" raising a SystemExit exception. This change wraps the main processing logic of Worker.workloop in a try/finally block which ensures that _ensure_messages_consumed is always called.